rename AuthHeader and update client deps
