+++ client/src/components/AuthHeader/AuthHeader.tsx
