+++ client/src/components/AuthHeader/useStyles.ts
